Linked the FilterOption with Notification file and removed the App.jsx code
