+++ package.json
@@ -38,6 +38,6 @@
     "postcss": "^8.4.49",
     "shadcn-ui": "^0.9.4",
     "tailwindcss": "^3.4.16",
-    "vite": "^6.0.1"
+    "vite": "^6.0.3"
   }
 }
