+++ src/Components/compIndex.js
@@ -1,4 +1,11 @@
 import Navbar from "./Navbar";
 import TopNavbar from "./TopNavbar";
+
 import Sidebar from "./Sidebar";
-export { Navbar, TopNavbar, Sidebar }
+import FilterOption from "./FilterOption"; // Ensure correct path 
+import Notification from "./Notification";
+
+export { Navbar, TopNavbar, Sidebar,FilterOption,Notification }
+
+
+
